requirements.txt
Django==1.7
Pillow==2.4.0
PyAudio==0.2.7
PyInstaller==2.1
PyYAML==3.11
Pygments==1.6
SQLAlchemy==0.9.4
Sphinx==1.2.2
Werkzeug==0.9.6
XlsxWriter==0.5.5
altgraph==0.12
argcomplete==0.6.7
astropy==0.3.2
atom==0.3.7
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.3.1
binstar==0.5.3
bitarray==0.8.1
blaze==0.5
blz==0.6.2
bokeh==0.4.4
boto==2.28.0
casuarius==1.1
cdecimal==2.3
chaco==4.4.1
chardet==2.2.1
colorama==0.2.7
conda==3.5.5
conda-build==1.3.5
configobj==5.0.5
copper==0.0.4
csvImporter==0.1.3.2
cubes==0.10.2
diff-match-patch==20121119
dj-database-url==0.3.0
django-adaptors==0.2.5
django-admin-import==0.3.0
django-admin-tools==0.5.2
django-appconf==0.6
django-csvimport==1.0
django-import-export==0.2.3
docutils==0.11
dowser==0.2
emport==1.0.0
enable==4.3.0
enaml==0.9.1
future==0.12.1
gevent==1.0.1
gevent-websocket==0.9.3
greenlet==0.4.2
grin==1.2.1
gunicorn==19.1.1
guppy==0.1.10
h5py==2.3.0
httplib2==0.9
infi.pyutils==1.0.8
infi.recipe.console-scripts==0.3.15
ipython==2.1.0
itsdangerous==0.24
jdcal==1.0
kiwisolver==0.1.2
llvmpy==0.12.6
lxml==3.3.5
macholib==1.7
matplotlib==1.3.1
mock==1.0.1
modulegraph==0.12
multipledispatch==0.4.3
networkx==1.8.1
nltk==2.0.4
nose==1.3.3
numba==0.13.2
numexpr==2.3.1
numpy==1.8.1
objgraph==1.8.1
openpyxl==1.8.5
pandas==0.14.0
patsy==0.2.1
pep8==1.5.6
ply==3.4
psutil==2.1.1
py==1.4.20
py2app==0.9
pycosat==0.6.1
pycparser==2.10
pycrypto==2.6.1
pycurl==7.19.3.1
pyface==4.4.0
pyflakes==0.8.1
pygraphviz==1.2
pyparsing==2.0.1
pytest==2.5.2
python-dateutil==1.5
pytz==2014.3
pyzmq==14.3.0
recordtype==1.1
redis==2.9.1
requests==2.3.0
rope==0.9.4
runipy==0.1.0
scikit-image==0.10.0
scikit-learn==0.13.1
scipy==0.14.0
six==1.6.1
spyder==2.3.0rc
statsmodels==0.5.0
sympy==0.7.5
tables==3.1.1
tablib==0.10.0
tornado==3.2.1
traits==4.4.0
traitsui==4.4.0
ujson==1.33
urlparse2==1.1.1
whitenoise==1.0.3
wsgiref==0.1.2
xlrd==0.9.3
xlwt==0.7.5
z3c.recipe.scripts==1.0.1
zc.buildout==2.2.1
zc.recipe.egg==2.0.1
psycopg2==2.5.3
